pyproject.toml
[project]
name = "pre-commit-fluxcd"
description = "Pre-Commit hooks for FluxCD GitOps repos"
dynamic = ["version"]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["pre-commit", "gitops", "fluxcd"]
requires-python = ">= 3.8"
dependencies = ["pyyaml==6.*"]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.urls]
Repository = "https://github.com/gabe565/pre-commit-fluxcd.git"
Issues = "https://github.com/gabe565/pre-commit-fluxcd/issues"

[project.scripts]
pre-commit-check-secrets-encrypted = "src.check_secrets_encrypted.__main__:main"
pre-commit-check-charts-pinned = "src.check_charts_pinned.__main__:main"
pre-commit-check-charts-support-renovate = "src.check_charts_support_renovate.__main__:main"
pre-commit-check-drift-detection-enabled = "src.check_drift_detection_enabled.__main__:main"
